Split the file into several.
